+++ app/utils/types.ts
@@ -765,6 +765,7 @@ export type UpdateInstructionPayload =
     | CooldownParameters
     | PoolParameters
     | TimeParameters
+    | TimeoutParameters
     | MinBlockTime
     | BlockEnergyLimit
     | FinalizationCommitteeParameters;
@@ -820,6 +821,7 @@ export enum UpdateType {
     PoolParameters,
     TimeParameters,
     UpdateMintDistributionV1,
+    TimeoutParameters,
     MinBlockTime,
     BlockEnergyLimit,
     FinalizationCommitteeParameters,
@@ -1110,6 +1112,12 @@ export function isBlockEnergyLimit(
     return UpdateType.BlockEnergyLimit === transaction.type;
 }
 
+export function isTimeoutParameters(
+    transaction: UpdateInstruction<UpdateInstructionPayload>
+): transaction is UpdateInstruction<TimeoutParameters> {
+    return UpdateType.TimeoutParameters === transaction.type;
+}
+
 export function isFinalizationCommitteeParameters(
     transaction: UpdateInstruction<UpdateInstructionPayload>
 ): transaction is UpdateInstruction<FinalizationCommitteeParameters> {
@@ -1239,6 +1247,12 @@ export interface FinalizationCommitteeParameters {
 
 export interface MinBlockTime {
     minBlockTime: Word64;
+}
+
+export interface TimeoutParameters {
+    timeoutBase: Word64;
+    timeoutIncrease: Fraction;
+    timeoutDecrease: Fraction;
 }
 
 export interface TimeParameters {