+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 -   Support for block energy limit chain update.
 -   Support for finalization committee parameters chain update.
 -   Support for minimum block time chain update.
+-   Support for timeout parameters chain update.
 
 ## 1.5.0